archive/release-v5.5/agents/mpa/base/functions/mpa_functions/requirements.txt
azure-functions
azure-identity
requests
python-dateutil
python-docx>=1.1.0
